scripts/requirements-generate-earnings-docs.txt
requests>=2.31.0
python-docx>=1.1.2
